little fixes

+++ inc/utils.h
@@ -2,6 +2,6 @@
 #define __UTILS_H__
 
 void printf (const char *format, ...);
-void itoa (char *buf, int base, int d);
+void itoa (char *buf, int base, long long d);
 
 #endif
